+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
 * [BREAKING] Replaced `messages` declaration with separate `success` and `error` calls
+* [BREAKING] Removed `rescues` method (use `error_from` for custom error messages; all exceptions now report to `on_exception` handlers)
 
 ## 0.1.0-alpha.2.6.1
 * [FEAT] Added `elapsed_time` and `outcome` methods to `Action::Result`